% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dim.r
\name{add_dim}
\alias{add_dim}
\title{Add array dimension}
\usage{
add_dim(x, where = NULL)
}
\arguments{
\item{x}{(required) an array}

\item{where}{vector of natural numbers specifying where the new dimension(s) 
should occur (see details). Defaults to last dimension.}
}
\value{
an array
}
\description{
Inserts a dimension into an array at a specified location
}
\details{
The function inserts a dimension into the array \code{x}
at each dimension given in \code{where}.
}
