% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_abun.r
\name{calc_abun}
\alias{calc_abun}
\title{Calculate global relative abundance}
\usage{
calc_abun(metacomm, N_S, only_species = F, ranked = F)
}
\arguments{
\item{metacomm}{(required) matrix of lists defining metacommunity
(as generated by \code{\link{populate_landscape}} or returned by
this function).}

\item{N_S}{(required) number of species expected in the metacommunity}

\item{only_species}{logical indicating whether the function should ignore
empty spaces. Defaults to \code{FALSE}}

\item{ranked}{logical indicating whether abundances should be returned in 
order of rank (\code{TRUE}) or in their original order (\code{FALSE}).
Default is \code{FALSE}.}
}
\value{
a vector of length \code{N_S} or, if \code{only_species=FALSE}, 
	a vector of length \code{N_S+1} where the first element corresponds to the number of
	empty spaces
}
\description{
Calculates the relative abundance or occupancy of each species in a metacommunity
}
\details{
The function counts the number of individuals of each species occuring
in a metacommunity and then divides by the total number of individuals.
If \code{only_species} is code{FALSE}, then empty spaces are included 
in the tally and the function returns the proportion of potential spaces
that each species occupies. Ranked abundances can be returned using the 
parameter \code{ranked}.
}
