% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rich.r
\name{calc_rich}
\alias{calc_rich}
\title{Richness profiles}
\usage{
calc_rich(locs = NULL, t_window = NULL, sim = NULL, N_S = NULL,
  abuns = NULL, agg_times = NULL, which_species = NULL)
}
\arguments{
\item{locs}{(required) either a two-column matrix of cell locations where 
species' abundances should be summed or a list of matrices designating
multiple spatial units where abundance should be summed (see details)}

\item{t_window}{(required) either a list containing \code{start}
and \code{stop} specifying that all collected timepoints in that interval
should be considered or an explicit vector of timepoints}

\item{sim}{(required) array of simulation results from a single 
simulation run (as returned by \code{\link{run_sim}})}

\item{N_S}{(required) number of species in the simulation}

\item{abuns}{array of abundance profiles (as returned by 
\code{\link{calc_abun_profile}})}

\item{agg_times}{either a single number specifing the number of timepoints 
that should be aggregated before calculating occupancy or a list of
vectors defining exactly which timepoints should be aggregated (see details).
 Defaults to no aggregation.}

\item{which_species}{a vector of integers indicating which species should be
considered. Defaults to all species.}
}
\value{
a matrix of species richness in each location at each timepoint 
(dimensions \code{[temporal unit, spatial unit]})
}
\description{
Calculate species richness through time for a simulation
}
\details{
The function calculates species richness for a set of locations
and timepoints over a given time interval. Users can either supply an 
array of abundance profiles (as returned by 
\code{\link{calc_abun_profile}}) or a set of spatial units
(\code{locs}), a time window (\code{t_window}), simulation
results (\code{sim}) and the number of species (\code{N_S}), 
from which abundance profiles can be calculated. These parameters
are explained in more detail in the documentation of 
\code{\link{calc_abun_profile}}. For each spatial and temporal unit in
the abundance profiles, the function calculates the number of species
present. Timepoints may be aggregated prior to this calculation using 
\code{agg_times}. This can be useful for evaluating the effects 
of sampling at different temporal scales on observed richness.
This aggregation is relative to the timepoints
stored in the abundance profiles array and not to the actual 
timesteps in the simulation. For example, specifying 
\code{agg_times=10} on abundance profiles containing
every other timestep will aggregate abundances from \code{T=1:20,
T=21:40,...}.
}
