% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctsim.r
\docType{package}
\name{CTSim}
\alias{CTSim}
\alias{CTSim-package}
\title{CTSim: a package for simulating core and transient species dynamics
in a spatially explicit metacommunity}
\description{
An explantion of the packages functionality can be found at this 
\href{https://github.com/hurlbertlab/core-transient-simulation/blob/master/README.md}{README}.
}
