% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disperse.r
\name{disperse}
\alias{disperse}
\title{Disperse individuals}
\usage{
disperse(x, y = NA, propagules, dim_land, d_rates, form = list(type =
  "gaussian"))
}
\arguments{
\item{x}{(required) x coordinate of the origin cell or vector of length
2 with x and y coordinates}

\item{y}{y coordinate of the origin cell}

\item{propagules}{(required) vector of integers representing dispersing 
individuals. Values indicate which species each individual belongs to.}

\item{dim_land}{(required) vector with the dimensions of the landscape: 
(x, y)}

\item{d_rates}{(required) vector of species dispersal rates from this cell
(based on its habitat type)}

\item{form}{list defining the dispersal kernel (see 
\code{\link{get_dispersal_vec}} for options). Defaults to Gaussian.}
}
\value{
a matrix whose columns give the x and y coordinates of the cells 
where dispersing individuals end up.
}
\description{
Disperses a set of individuals from a given cell and determines the locations 
of the new cells that each lands in.
}
\details{
For each dispersing individual, the function choses a random location 
within the cell of origin and then adds a dispersal vector, which is 
determined by the function \code{\link{get_dispersal_vec}}. The distance 
traveled is determined stochastically based on the dispersal kernel
defined in \code{form} and the species-specific dispersal rate defined in
the \code{d_rates} vector. Individuals are allowed to disperse off of the
landscape and the function may return cell locations outside of the 
intervals \code{[1,x] and [1,y]}.
}
\seealso{
\code{\link{get_dispersal_vec}} for calculation of dispersal vectors 
	and for how to specify dispersal kernels.
}
