% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/establish.r
\name{establish}
\alias{establish}
\title{Establishment}
\usage{
establish(comm, propagules, r_rates, m = 0, gsad = NULL)
}
\arguments{
\item{comm}{(required) vector (or list) of species present- may contain 
zeros for an empty spaces}

\item{propagules}{(required) vector of integers representing pool of 
potential recruits. Values indicate species identity of each individual.}

\item{r_rates}{(required) vector of species' recruitment rates for this cell
(based on its habitat type). See details.}

\item{m}{immigration rate. Defaults to 0.}

\item{gsad}{vector defining the global relative abundance of each species.
    Must be in the same order as \code{species}. Defaults to same abundance 
for all species. See details.}
}
\value{
a vector of the same length as \code{comm} defining the new community
}
\description{
Determines which individuals establish in the empty spaces in a community, 
given a pool of potenial recruits and probability that migrants will arrive
from outside the metacommunity.
}
\details{
For each empty space in the community \code{comm} (designated by a 0), 
the function first attempts to fill the space with a migrant from
outside the metacommunity with probability \code{m}. If successful, the species
identity is probabilistically chosen according to the relative abundances
provided in \code{gsad}. Spaces remaining empty after immigration are filled
stochastically from the pool of potential recruits provided in 
\code{propagules}. Individuals are chosen with equal probability.
Species recruitment rates define the probability that an individual 
selected for an empty space will actually establish.
}
