% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dispersal_vec.r
\name{get_dispersal_vec}
\alias{get_dispersal_vec}
\title{Calculate dispersal vector}
\usage{
get_dispersal_vec(d, form = list(type = "gaussian"), N = 1)
}
\arguments{
\item{d}{(required) numeric controlling dispersal distance (see details).}

\item{form}{list defining the dispersal kernel (see details).
Minimally contains an argument \code{type} defining the form of the 
dispersal kernel. Defaults to Gaussian.}

\item{N}{number of vectors to generate}
}
\value{
a matrix of dispersal vectors. First column gives the magnitude 
and the second column gives the direction in radians, counter-clockwise
from East.
}
\description{
Stochastically generates dispersal vectors given a dispersal kernel 
and expected dispersal distance.
}
\details{
Only implements isotropic dispersal (no directional bias). Three types 
of dispersal kernels are currently implemented and should be specified 
with the \code{type} argument in the list which is passed as the parameter 
\code{form}. These are:
\describe{
	\item{\code{'gaussian'}}{Gaussian kernel- dispersal distance sampled from
		a half-normal distribution with mean \code{d}.}
	\item{\code{'adjacent'}}{Adjacent cell dispersal (rook's move)- \code{d} 
		is the probability that an individual will move. An
		additional argument \code{moves} must be passed in \code{form} 
		giving the number of steps an individual should attempt to take.}
	\item{\code{'uniform'}}{Uniform dispersal- individual has equal 
		probability of occurring anywhere within a circle with radius 
		\code{d}. Sufficiently large \code{d} allows for unlimited dispersal.}
}
}
\seealso{
\code{\link{disperse}} for dispersal of propagules with different 
expected dispersal distances.
}
