% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_parms.r
\name{write_parms}
\alias{write_parms}
\title{Write a parameter file}
\usage{
write_parms(parm_list, f)
}
\arguments{
\item{parm_list}{(required) list of parameters, as generated by 
\code{\link{make_parmlist}}}

\item{f}{(required) file path}
}
\value{
nothing

export
}
\description{
Writes a list of parameters to a file
}
\details{
This function creates a parameter file that can be read by 
\code{\link{run_sim_P}}. It should be used in conjection with 
\code{\link{make_parmlist}}.
}
